requirements.txt
discord.py[voice]
emc-info == 0.8
firebase_admin
youtube_dl
PyNaCl
